+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Correctly apply optimistic concurrency controls to individual storage objects under high write contention.
 - Time spent metrics are now correctly reported in milliseconds.
 - Password minimum length error message now correctly reflects the constraint.
+- Explicitly set response Content-Type header in successful HTTP RPC responses.
 
 ## [2.7.0] - 2019-09-11
 ### Added